Version bump and directory restructure
