+++ src/mappings/v2/token-marketplace-v2-mapping.ts
@@ -8,29 +8,29 @@ import {
     BidPlaced,
     BidRejected,
     BidWithdrawn
-} from "../../generated/TokenMarketplace/TokenMarketplace";
+} from "../../../generated/TokenMarketplace/TokenMarketplace";
 
-import {TokenOffer} from "../../generated/schema";
+import {TokenOffer} from "../../../generated/schema";
 
-import {toEther} from "../utils/utils";
+import {toEther} from "../../utils/utils";
 
-import {getKnownOriginV2ForAddress} from "../utils/KODAV2AddressLookup";
+import {getKnownOriginV2ForAddress} from "./KODAV2AddressLookup";
 
-import {TokenDeListed, TokenListed, TokenPurchased} from "../../generated/TokenMarketplaceV2/TokenMarketplaceV2";
-import {ONE, ZERO, ZERO_BIG_DECIMAL} from "../utils/constants";
+import {TokenDeListed, TokenListed, TokenPurchased} from "../../../generated/TokenMarketplaceV2/TokenMarketplaceV2";
+import {ONE, ZERO, ZERO_BIG_DECIMAL} from "../../utils/constants";
 
-import * as KodaVersions from "../utils/KodaVersions";
-import * as SaleTypes from "../utils/SaleTypes";
+import * as KodaVersions from "../../utils/KodaVersions";
+import * as SaleTypes from "../../utils/SaleTypes";
 
-import * as editionService from "../services/Edition.service";
-import * as tokenEventFactory from "../services/TokenEvent.factory";
-import * as collectorService from "../services/Collector.service";
-import * as listedTokenService from "../services/ListedToken.service";
-import * as tokenService from "../services/Token.service";
-import * as dayService from "../services/Day.service";
-import * as offerService from "../services/Offers.service";
-import * as activityEventService from "../services/ActivityEvent.service";
-import * as artistService from "../services/Artist.service";
+import * as editionService from "../../services/Edition.service";
+import * as tokenEventFactory from "../../services/TokenEvent.factory";
+import * as collectorService from "../../services/Collector.service";
+import * as listedTokenService from "../../services/ListedToken.service";
+import * as tokenService from "../../services/Token.service";
+import * as dayService from "../../services/Day.service";
+import * as offerService from "../../services/Offers.service";
+import * as activityEventService from "../../services/ActivityEvent.service";
+import * as artistService from "../../services/Artist.service";
 
 export function handleAuctionEnabled(event: AuctionEnabled): void {
     /*
